+++ src/Interfaces/WebhookServiceInterface.php
@@ -65,7 +65,7 @@ interface WebhookServiceInterface
      * @throws UnzerApiException An UnzerApiException is thrown if there is an error returned on API-request.
      * @throws RuntimeException  A RuntimeException is thrown when there is an error while using the SDK.
      */
-    public function updateWebhook($webhook): Webhook;
+    public function updateWebhook(Webhook $webhook): Webhook;
 
     /**
      * Deletes the given Webhook resource.
@@ -120,5 +120,5 @@ interface WebhookServiceInterface
      * @throws UnzerApiException An UnzerApiException is thrown if there is an error returned on API-request.
      * @throws RuntimeException  A RuntimeException is thrown when there is an error while using the SDK.
      */
-    public function fetchResourceFromEvent($eventJson = null): AbstractUnzerResource;
+    public function fetchResourceFromEvent(string $eventJson = null): AbstractUnzerResource;
 }