+++ src/Resources/EmbeddedResources/CompanyInfo.php
@@ -62,7 +62,7 @@ class CompanyInfo extends AbstractUnzerResource
      *
      * @return CompanyInfo
      */
-    public function setRegistrationType($registrationType): CompanyInfo
+    public function setRegistrationType(?string $registrationType): CompanyInfo
     {
         $this->registrationType = $this->removeRestrictedSymbols($registrationType);
         return $this;
@@ -81,7 +81,7 @@ class CompanyInfo extends AbstractUnzerResource
      *
      * @return CompanyInfo
      */
-    public function setCommercialRegisterNumber($commercialRegisterNumber): CompanyInfo
+    public function setCommercialRegisterNumber(?string $commercialRegisterNumber): CompanyInfo
     {
         $this->commercialRegisterNumber = empty($commercialRegisterNumber) ?
             $commercialRegisterNumber : $this->removeRestrictedSymbols($commercialRegisterNumber);
@@ -101,7 +101,7 @@ class CompanyInfo extends AbstractUnzerResource
      *
      * @return CompanyInfo
      */
-    public function setFunction($function): CompanyInfo
+    public function setFunction(?string $function): CompanyInfo
     {
         $this->function = $this->removeRestrictedSymbols($function);
         return $this;
@@ -193,7 +193,7 @@ class CompanyInfo extends AbstractUnzerResource
      *
      * @return mixed
      */
-    private function removeRestrictedSymbols($value)
+    private function removeRestrictedSymbols(?string $value)
     {
         if (!is_string($value)) {
             return $value;