+++ src/Services/PaymentService.php
@@ -92,7 +92,7 @@ class PaymentService implements PaymentServiceInterface
 
     //<editor-fold desc="Authorize transaction">
 
-    public function performAuthorization(Authorization $authorization, $paymentType, $customer = null, $metadata = null, $basket = null): Authorization
+    public function performAuthorization(Authorization $authorization, $paymentType, $customer = null, $metadata = null, Basket $basket = null): Authorization
     {
         $payment = $this->createPayment($paymentType);
         $paymentType = $payment->getPaymentType();
@@ -162,7 +162,7 @@ class PaymentService implements PaymentServiceInterface
     /**
      * {@inheritDoc}
      */
-    public function performCharge(Charge $charge, $paymentType, $customer = null, $metadata = null, $basket = null): Charge
+    public function performCharge(Charge $charge, $paymentType, $customer = null, Metadata $metadata = null, Basket $basket = null): Charge
     {
         $payment     = $this->createPayment($paymentType);
         $paymentType = $payment->getPaymentType();
@@ -280,16 +280,16 @@ class PaymentService implements PaymentServiceInterface
      * {@inheritDoc}
      */
     public function payout(
-        $amount,
-        $currency,
+        float    $amount,
+        string   $currency,
         $paymentType,
-        $returnUrl,
+        string   $returnUrl,
         $customer = null,
-        $orderId = null,
-        $metadata = null,
-        $basket = null,
-        $invoiceId = null,
-        $referenceText = null
+        string   $orderId = null,
+        Metadata $metadata = null,
+        Basket   $basket = null,
+        string   $invoiceId = null,
+        string $referenceText = null
     ): Payout {
         $payment = $this->createPayment($paymentType);
         $payout = (new Payout($amount, $currency, $returnUrl))
@@ -356,9 +356,9 @@ class PaymentService implements PaymentServiceInterface
      * {@inheritDoc}
      */
     public function fetchInstallmentPlans(
-        $amount,
-        $currency,
-        $effectiveInterest,
+        float    $amount,
+        string   $currency,
+        float    $effectiveInterest,
         DateTime $orderDate = null
     ): InstalmentPlans {
         $ins   = (new InstallmentSecured(null, null, null))->setParentResource($this->unzer);
@@ -392,10 +392,10 @@ class PaymentService implements PaymentServiceInterface
      * @throws RuntimeException  A RuntimeException is thrown when there is an error while using the SDK.
      */
     private function initPayPage(
-        Paypage $paypage,
-        $action,
+        Paypage  $paypage,
+        string   $action,
         Customer $customer = null,
-        Basket $basket = null,
+        Basket   $basket = null,
         Metadata $metadata = null
     ): Paypage {
         $paypage->setAction($action)->setParentResource($this->unzer);